Count the discarded HEAD body in an Int64
NullIO counts the body a HEAD request generates so it can set
Content-Length, and the counter was an Int32. Crystal checks integer
overflow, so a body of 2 GiB or more raised OverflowError, which
ExceptionHandler turned into a 500 while the matching GET answered 200.
The threshold was exactly Int32::MAX.

A route generating a body over that size also reported the partial count
as its Content-Length, since NullIO sets the header from the counter when
the handler has not set one itself.

HTTP::Server::Response#content_length already returns Int64?, so the
counter was the only Int32 left in the path.

Fixes #803

+++ src/kemal/head_request_handler.cr
@@ -8,13 +8,13 @@ module Kemal
 
     private class NullIO < IO
       @original_output : IO
-      @out_count : Int32
+      @out_count : Int64
       @response : HTTP::Server::Response
 
       def initialize(@response)
         @closed = false
         @original_output = @response.output
-        @out_count = 0
+        @out_count = 0_i64
       end
 
       def read(slice : Bytes)
